+++ src/redshift/redshift/mod.rs
@@ -8,4 +8,3 @@ pub mod generators;
 pub mod prover;
 pub mod verifier;
 pub mod tests;
-//pub mod verifying_circuit;
