--- src/redshift/redshift/verifying_circuit/mod.rs
@@ -1,5 +1,0 @@
-pub mod common;
-pub mod rescue_gadget;
-pub mod transplilation_check;
-pub mod boolean;
-//pub mod merklee_tree_gadget;
